requirements.txt
pydantic>=2.6.0
openai>=1.30.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
typing-extensions>=4.9.0
google-genai>=1.0.0